+++ extensions/todoist/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Todoist Changelog
 
-## [Add Support for Task Deadlines] - {PR_MERGE_DATE}
+## [Add Complete Task Shortcut Preferences] - 2025-01-31
+
+- Adjusting shortcuts for the following task actions:
+  - Complete Task - Now `⌘` + `Enter` (previously `shift` + `⌘` + `E`)
+  - Open Task in Todoist - Now `⌘` + `O` (previously `⌘` + `Enter`)
+
+## [Add Support for Task Deadlines] - 2025-01-31
 
 It's now possible to specify a deadline when using the `Create Task` command. Deadlines are displayed in the task list and task details views.